Remove Object#assert_respond_to!
In many cases it is likely preferable to simply call the method in
question and let Ruby raise an appropriate exception.

In cases where an assertion is necessary, the same check can be
performed with `assert!{|object| object.respond_to?(method_name) }`, at
the cost of a less-descriptive error message.